requirements.txt
api==0.0.7
seaborn==0.12.2
numpy==1.24.1
matplotlib==3.6.3
pandas==1.5.3
asgiref==3.5.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
environ==1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
mysql-connector-python==8.0.31
nose==1.3.7
packaging==22.0
protobuf==3.20.1
pyrsistent==0.19.2
pytz==2022.6
#pypiwin32==223
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
websocket-client==1.4.2
uwsgi